Document post-approve landing publication flow and responsibilities
