+++ backend/ads-service/src/main/java/com/marketinghub/geralanding/DesignPresetProvisionalHtmlAssembler.java
@@ -1,10 +1,6 @@
 package com.marketinghub.geralanding;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-import org.jsoup.parser.Parser;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
@@ -42,7 +38,6 @@ public class DesignPresetProvisionalHtmlAssembler {
                     objectMapper.writeValueAsString(copyPayload),
                     imagePlanningJson,
                     designPresetOutputJson);
-            html = appendBehaviorTrackingAttributesAndScript(html);
             return appendJobIdCommentBeforeHead(html, jobId);
         } catch (Exception e) {
             throw new IllegalArgumentException("Falha ao montar HTML provisório da fase landing-page-design-preset", e);
@@ -70,84 +65,4 @@ public class DesignPresetProvisionalHtmlAssembler {
         return html.substring(0, headIndex) + comment + html.substring(headIndex);
     }
 
-    private String appendBehaviorTrackingAttributesAndScript(String html) {
-        if (!StringUtils.hasText(html) || html.contains("data-mh-funnel-tracking")) {
-            return html;
-        }
-        Document document = Jsoup.parse(html, "", Parser.htmlParser());
-        document.outputSettings().prettyPrint(false);
-
-        for (Element section : document.select("section[data-section-id], section[id], [data-section-id]")) {
-            String sectionId = section.hasAttr("data-section-id") ? section.attr("data-section-id") : section.id();
-            if (!StringUtils.hasText(sectionId)) {
-                continue;
-            }
-            section.attr("data-track-section", sectionId.trim());
-        }
-
-        String script = """
-                <script data-mh-funnel-tracking="true">
-                (function(){
-                  if (window.__mhFunnelTrackingInstalled) return;
-                  window.__mhFunnelTrackingInstalled = true;
-                  var debugPrefix = '[MH funnel tracking]';
-                  window.dataLayer = window.dataLayer || [];
-                  function emit(name, payload){
-                    var eventPayload = Object.assign({event:name, source:'landing-page-design-preset'}, payload||{});
-                    console.debug(debugPrefix, 'emit', eventPayload);
-                    window.dataLayer.push(eventPayload);
-                  }
-                  console.debug(debugPrefix, 'bootstrap');
-                  emit('page_view', {ts: Date.now()});
-                  var sections = Array.prototype.slice.call(document.querySelectorAll('[data-track-section]'));
-                  console.debug(debugPrefix, 'sections-found', {count: sections.length});
-                  var stats = {};
-                  sections.forEach(function(node){
-                    var id = node.getAttribute('data-track-section');
-                    stats[id] = {visibleSince:null, elapsedMs:0};
-                  });
-                  function flushSection(id, reason){
-                    var s = stats[id];
-                    if (!s || s.visibleSince === null) return;
-                    s.elapsedMs += Date.now() - s.visibleSince;
-                    s.visibleSince = null;
-                    console.debug(debugPrefix, 'section-flush', {sectionId:id, elapsedMs:s.elapsedMs, reason: reason || 'hidden'});
-                    emit('section_view_time', {sectionId:id, elapsedMs:s.elapsedMs, reason: reason || 'hidden'});
-                  }
-                  var observer = new IntersectionObserver(function(entries){
-                    entries.forEach(function(entry){
-                      var id = entry.target.getAttribute('data-track-section');
-                      if (!id || !stats[id]) return;
-                      if (entry.isIntersecting && entry.intersectionRatio >= 0.5) {
-                        if (stats[id].visibleSince === null) {
-                          stats[id].visibleSince = Date.now();
-                          console.debug(debugPrefix, 'section-visible', {sectionId:id, ts:stats[id].visibleSince});
-                          emit('section_view_start', {sectionId:id});
-                        }
-                      } else {
-                        flushSection(id, 'intersection-change');
-                      }
-                    });
-                  }, {threshold:[0,0.5,1]});
-                  sections.forEach(function(node){ observer.observe(node); });
-                  document.addEventListener('visibilitychange', function(){
-                    if (document.hidden) {
-                      console.debug(debugPrefix, 'visibility-hidden');
-                      Object.keys(stats).forEach(function(id){ flushSection(id, 'tab-hidden'); });
-                    }
-                  });
-                  window.addEventListener('beforeunload', function(){
-                    console.debug(debugPrefix, 'beforeunload-flush');
-                    Object.keys(stats).forEach(function(id){ flushSection(id, 'before-unload'); });
-                  });
-                })();
-                </script>
-                """;
-        if (document.head() != null) {
-            document.head().append(script);
-        } else {
-            document.prepend(script);
-        }
-        return document.outerHtml();
-    }
 }
