+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/DesignPresetProvisionalHtmlAssemblerTest.java
@@ -3,6 +3,7 @@ package com.marketinghub.geralanding;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -11,7 +12,7 @@ import static org.mockito.Mockito.when;
 class DesignPresetProvisionalHtmlAssemblerTest {
 
     @Test
-    void shouldInjectBehaviorTrackingIntoDesignPresetProvisionalHtml() {
+    void shouldAssembleHtmlWithoutInjectingBehaviorTrackingScript() {
         DesignPresetProvisionalHtmlProcessor processor = mock(DesignPresetProvisionalHtmlProcessor.class);
         when(processor.process(anyString(), anyString(), anyString(), anyString()))
                 .thenReturn("<html><head><title>x</title></head><body><section data-section-id='hero'></section></body></html>");
@@ -24,9 +25,10 @@ class DesignPresetProvisionalHtmlAssemblerTest {
                 "{\"landingPageDesignPreset\":{\"sectionPresets\":[]}}",
                 "job-123");
 
-        assertTrue(html.contains("data-mh-funnel-tracking=\"true\""));
-        assertTrue(html.contains("page_view"));
-        assertTrue(html.contains("section_view_time"));
-        assertTrue(html.contains("data-track-section=\"hero\""));
+        assertFalse(html.contains("data-mh-funnel-tracking=\"true\""));
+        assertFalse(html.contains("page_view"));
+        assertFalse(html.contains("section_view_time"));
+        assertFalse(html.contains("data-track-section=\"hero\""));
+        assertTrue(html.contains("<!-- jobId = job-123 -->"));
     }
 }
