+++ docs/canonical/procedimento-experimento-canon.v1.md
@@ -72,7 +72,7 @@ Regras complementares:
 | `LANDING_PAGE_WIREFRAME` | `WireframeProvisionalHtmlAssembler` | `gera_landing_stage_execution.provisional_html` |
 | `LANDING_PAGE_COPY` | `CopyProvisionalHtmlAssembler` | `gera_landing_stage_execution.provisional_html` |
 | `LANDING_PAGE_IMAGE_PLANNING` | Não há assembler exclusivo da etapa; usa `CopyProvisionalHtmlAssembler.assembleComplete(...)` + `LandingPageImageInjector.injectImages(...)`. | `gera_landing_stage_execution.provisional_html` e `experiment.landing_page_html` (quando `provisionalHtml` existe na execução). |
-| `LANDING_PAGE_DESIGN_PRESET` | `DesignPresetProvisionalHtmlAssembler` + `LandingPageImageInjector.injectImages(...)` | `gera_landing_stage_execution.provisional_html` e `experiment.landing_page_html` |
+| `LANDING_PAGE_DESIGN_PRESET` | `DesignPresetProvisionalHtmlAssembler` + `LandingPageImageInjector.injectImages(...)` | `gera_landing_stage_execution.provisional_html`, `experiment.landing_page_design_preset` (HTML consolidado da etapa) e `experiment.landing_page_html` |
 
 ## 6. Geração e aprovação de anúncios
 
@@ -105,6 +105,22 @@ Regras obrigatórias do modo Flex no Gera Landing:
 ## 9. Publicação da landing
 
 A publicação do HTML final da landing ocorre no Lead Portal, com integração feita pelo fluxo de backend/worker.
+
+### 9.1 O que acontece depois de clicar em "Aprovar e publicar landing"
+
+Classe responsável no backend: `GeraLandingStageExecutionService` (método `approveAndPublishLanding`).
+
+Fluxo obrigatório executado após a aprovação:
+1. carregar o experimento e resolver o HTML base para publicação (priorizando `experiment.landing_page_design_preset`; fallback legado para `experiment.landing_page_html`);
+2. injetar instrumentação de tracking comportamental no HTML publicado (`data-track-section` + script `data-mh-funnel-tracking`);
+3. injetar controles de funil (`data-mh-funnel-controls`);
+4. resolver `facebookPixelId` do nicho do experimento e injetar snippet do Facebook Pixel quando elegível;
+5. publicar o flow no Lead Portal via `PUT /api/flows/{slug}` com payload contendo `slug`, `name`, `description` e `customFormHtml`;
+6. resolver URLs finais de publicação (`iframe` e `standalone`) e persistir no experimento a `follow_up_action_url`.
+
+Regras adicionais:
+- a injeção de tracking deve ser idempotente (não duplicar quando já existir no HTML);
+- falhas de contrato na publicação para Lead Portal devem ser tratadas pela exception canônica de violação de contrato do GeraLanding.
 
 ## 10. Custos e mensuração por experimento
